strorage: browser session and user password repositories

+++ crates/templates/src/context.rs
@@ -532,14 +532,14 @@ where {
 /// Context used by the `account/index.html` template
 #[derive(Serialize)]
 pub struct AccountContext {
-    active_sessions: i64,
+    active_sessions: usize,
     emails: Vec<UserEmail>,
 }
 
 impl AccountContext {
     /// Constructs a context for the "my account" page
     #[must_use]
-    pub fn new(active_sessions: i64, emails: Vec<UserEmail>) -> Self {
+    pub fn new(active_sessions: usize, emails: Vec<UserEmail>) -> Self {
         Self {
             active_sessions,
             emails,
